Rename module names
Problem: module names are prefixed with `Crv` which does not suit the
new project naming.

Solution: following the new `xrefcheck` name, rename modules so that
they start from `Xrefcheck`.
